# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PCCScribe API - Clinical note summarization platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface Patient {
  id: number;
  name: string;
  age: number;
  facilityName: string;
  unit: string;
  mrn?: string | null;
  admissionDate?: string | null;
  primaryDiagnosis?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePatientBody {
  name: string;
  age: number;
  facilityName: string;
  unit: string;
  mrn?: string | null;
  admissionDate?: string | null;
  primaryDiagnosis?: string | null;
}

export type ClinicalNoteNoteType =
  (typeof ClinicalNoteNoteType)[keyof typeof ClinicalNoteNoteType];

export const ClinicalNoteNoteType = {
  progress_notes: "progress_notes",
  physician_orders: "physician_orders",
  mds_assessment: "mds_assessment",
  care_plan: "care_plan",
  mar: "mar",
  nursing_notes: "nursing_notes",
  therapy_notes: "therapy_notes",
  dietary_notes: "dietary_notes",
  social_work_notes: "social_work_notes",
  other: "other",
} as const;

export interface ClinicalNote {
  id: number;
  patientId: number;
  noteType: ClinicalNoteNoteType;
  noteDate: string;
  author?: string | null;
  content: string;
  sourceUrl?: string | null;
  createdAt: string;
}

export type IngestNotesBodyNotesItemNoteType =
  (typeof IngestNotesBodyNotesItemNoteType)[keyof typeof IngestNotesBodyNotesItemNoteType];

export const IngestNotesBodyNotesItemNoteType = {
  progress_notes: "progress_notes",
  physician_orders: "physician_orders",
  mds_assessment: "mds_assessment",
  care_plan: "care_plan",
  mar: "mar",
  nursing_notes: "nursing_notes",
  therapy_notes: "therapy_notes",
  dietary_notes: "dietary_notes",
  social_work_notes: "social_work_notes",
  other: "other",
} as const;

export type IngestNotesBodyNotesItem = {
  noteType: IngestNotesBodyNotesItemNoteType;
  noteDate: string;
  author?: string | null;
  content: string;
  sourceUrl?: string | null;
};

export type IngestNotesBodySource =
  (typeof IngestNotesBodySource)[keyof typeof IngestNotesBodySource];

export const IngestNotesBodySource = {
  extension: "extension",
  manual: "manual",
} as const;

export interface IngestNotesBody {
  notes: IngestNotesBodyNotesItem[];
  source?: IngestNotesBodySource;
}

export type SummaryStatus = (typeof SummaryStatus)[keyof typeof SummaryStatus];

export const SummaryStatus = {
  pending: "pending",
  generating: "generating",
  completed: "completed",
  failed: "failed",
} as const;

export type SummaryConfidence =
  | (typeof SummaryConfidence)[keyof typeof SummaryConfidence]
  | null;

export const SummaryConfidence = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type SummarySoapSummarySubjective = {
  narrative: string;
  patientReportedSymptoms: string[];
  familyConcernsDocumented: string[];
  painReported: boolean;
  painScaleReported?: string | null;
  moodAffect?: string | null;
  dataSufficiency: string;
};

export type SummarySoapSummaryObjectiveKeyFindingsItem = {
  finding: string;
  source: string;
};

export type SummarySoapSummaryObjective = {
  narrative: string;
  keyFindings: SummarySoapSummaryObjectiveKeyFindingsItem[];
  vitalsTrend?: string | null;
  functionalStatus?: string | null;
  dataSufficiency: string;
};

export type SummarySoapSummaryAssessment = {
  narrative: string;
  activeDiagnoses: string[];
  clinicalTrajectory?: string | null;
  riskFlags: string[];
  dataSufficiency: string;
};

export type SummarySoapSummaryPlan = {
  narrative: string;
  activeOrders: string[];
  pendingActions: string[];
  followUpItems: string[];
  dataSufficiency: string;
};

export type SummarySoapSummary = {
  subjective: SummarySoapSummarySubjective;
  objective: SummarySoapSummaryObjective;
  assessment: SummarySoapSummaryAssessment;
  plan: SummarySoapSummaryPlan;
} | null;

export type SummaryPerNoteTypeSummaries = {
  [key: string]: {
    confidence: string;
    summary: string;
    notesCount: number;
    keyPoints: string[];
  };
} | null;

export type SummaryKeyClinicalEventsItemSignificance =
  (typeof SummaryKeyClinicalEventsItemSignificance)[keyof typeof SummaryKeyClinicalEventsItemSignificance];

export const SummaryKeyClinicalEventsItemSignificance = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type SummaryKeyClinicalEventsItem = {
  date: string;
  event: string;
  significance: SummaryKeyClinicalEventsItemSignificance;
};

export interface Summary {
  id: number;
  patientId: number;
  dateFrom: string;
  dateTo: string;
  noteTypesIncluded: string[];
  notesCount: number;
  status: SummaryStatus;
  confidence?: SummaryConfidence;
  oneLiner?: string | null;
  soapSummary?: SummarySoapSummary;
  perNoteTypeSummaries?: SummaryPerNoteTypeSummaries;
  keyClinicalEvents?: SummaryKeyClinicalEventsItem[] | null;
  documentationGaps?: string[] | null;
  errorMessage?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface GenerateSummaryBody {
  /** Start date for notes to include (YYYY-MM-DD) */
  dateFrom: string;
  /** End date for notes to include (YYYY-MM-DD) */
  dateTo: string;
  /** Filter to specific note types. Empty means all. */
  noteTypes?: string[];
}

export type CreatePatientNotes201 = {
  inserted: number;
  noteIds: number[];
};