# lib/api-zod/src/generated/types/clinicalNote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PCCScribe API - Clinical note summarization platform
 * OpenAPI spec version: 0.1.0
 */
import type { ClinicalNoteNoteType } from "./clinicalNoteNoteType";

export interface ClinicalNote {
  id: number;
  patientId: number;
  noteType: ClinicalNoteNoteType;
  noteDate: string;
  author?: string | null;
  content: string;
  sourceUrl?: string | null;
  createdAt: string;
}