# lib/api-zod/src/generated/types/createPatientBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PCCScribe API - Clinical note summarization platform
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePatientBody {
  name: string;
  age: number;
  facilityName: string;
  unit: string;
  mrn?: string | null;
  admissionDate?: string | null;
  primaryDiagnosis?: string | null;
}