# lib/api-zod/src/generated/types/generateSummaryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PCCScribe API - Clinical note summarization platform
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateSummaryBody {
  /** Start date for notes to include (YYYY-MM-DD) */
  dateFrom: string;
  /** End date for notes to include (YYYY-MM-DD) */
  dateTo: string;
  /** Filter to specific note types. Empty means all. */
  noteTypes?: string[];
}